Add env
Reorg paths
