--- settings.json
@@ -1,6 +1,0 @@
-{
-  "officeBGExe":"LIBRE_OFFICE", 
-  "libreOfficePath":"C:\\Program Files\\LibreOffice\\program\\soffice.exe",
-  "inputPath":"C:\\Users\\cyril\\OneDrive\\Documents\\Solaris\\nextcloud\\Solaris\\temp",
-  "outputPath":"C:\\Users\\cyril\\OneDrive\\Documents\\Solaris\\nextcloud\\Solaris\\temp"
-}
