+++ src/settings.json
@@ -1,0 +1,6 @@
+{
+  "officeBGExe":"LIBRE_OFFICE", 
+  "libreOfficePath":"C:\\Program Files\\LibreOffice\\program\\soffice.exe",
+  "inputPath":"C:\\Workspace\\Python\\ConvertDirectoryToPDF\\tests\\sources_folder",
+  "outputPath":"C:\\temp"
+}
